+++ examples/backend-adapters/client-vanilla/src/uploadthing.ts
@@ -7,7 +7,7 @@ import { EndpointMetadata } from "uploadthing/types";
 
 import type { OurFileRouter } from "../../server/src/router";
 
-const BASE_URL = "http://localhost:3003";
+const BASE_URL = "http://localhost:3000";
 
 export const { uploadFiles, createUpload } = genUploader<OurFileRouter>({
   url: BASE_URL,