+++ examples/minimal-tanstack-start/app/client.tsx
@@ -1,3 +1,4 @@
+/// <reference types="vinxi/types/client" />
 import { StartClient } from "@tanstack/start";
 import { hydrateRoot } from "react-dom/client";