+++ packages/expo/src/index.ts
@@ -3,7 +3,7 @@ import Constants from "expo-constants";
 import { generateReactHelpers } from "@uploadthing/react/native";
 import { warnIfInvalidPeerDependency } from "@uploadthing/shared";
 import { version as uploadthingClientVersion } from "uploadthing/client";
-import type { FileRouter } from "uploadthing/internal/types";
+import type { FileRouter } from "uploadthing/types";
 
 import { peerDependencies } from "../package.json";
 import { GENERATE_useDocumentUploader } from "./document-picker";
@@ -38,7 +38,7 @@ export const generateReactNativeHelpers = <TRouter extends FileRouter>(
       initOpts?.url ?? "/api/uploadthing",
       typeof window.location !== "undefined"
         ? window.location.origin
-        : process.env.EXPO_PUBLIC_SERVER_ORIGIN ?? `http://${debuggerHost}`,
+        : (process.env.EXPO_PUBLIC_SERVER_ORIGIN ?? `http://${debuggerHost}`),
     );
   } catch (e) {
     // Can't throw since window.location is undefined in Metro pass