+++ packages/nuxt/playground/nuxt.config.ts
@@ -1,5 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   devtools: { enabled: true },
-  modules: ["../src/module"],
+  modules: ["../src/module", "@nuxtjs/tailwindcss"],
+  telemetry: false,
 });