+++ packages/uploadthing/src/express.ts
@@ -10,14 +10,14 @@ import type { Json } from "@uploadthing/shared";
 
 import { makeAdapterHandler } from "./internal/handler";
 import { getPostBody, toWebRequest } from "./internal/to-web-request";
-import type { FileRouter, RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
+import type { FileRouter, RouteHandlerOptions } from "./types";
 
 export { UTFiles } from "./internal/types";
 export type { FileRouter };
 
-type MiddlewareArgs = {
+type AdapterArgs = {
   req: ExpressRequest;
   res: ExpressResponse;
   event: undefined;
@@ -25,7 +25,7 @@ type MiddlewareArgs = {
 
 export const createUploadthing = <TErrorShape extends Json>(
   opts?: CreateBuilderOptions<TErrorShape>,
-) => createBuilder<MiddlewareArgs, TErrorShape>(opts);
+) => createBuilder<AdapterArgs, TErrorShape>(opts);
 
 export const createRouteHandler = <TRouter extends FileRouter>(
   opts: RouteHandlerOptions<TRouter>,