+++ packages/uploadthing/src/sdk/utils.ts
@@ -1,14 +1,10 @@
-import {
-  HttpClient,
-  HttpClientRequest,
-  HttpClientResponse,
-} from "@effect/platform";
+import { HttpClient, HttpClientRequest } from "@effect/platform";
 import * as Effect from "effect/Effect";
+import * as Predicate from "effect/Predicate";
 
 import {
   generateKey,
   generateSignedURL,
-  isObject,
   UploadThingError,
 } from "@uploadthing/shared";
 import type {
@@ -85,7 +81,7 @@ export const downloadFiles = (
     urls,
     (_url, idx) =>
       Effect.gen(function* () {
-        let url = isObject(_url) ? _url.url : _url;
+        let url = Predicate.isRecord(_url) ? _url.url : _url;
         if (typeof url === "string") {
           // since dataurls will result in name being too long, tell the user
           // to use uploadFiles instead.
@@ -105,12 +101,15 @@ export const downloadFiles = (
         const {
           name = url.pathname.split("/").pop() ?? "unknown-filename",
           customId = undefined,
-        } = isObject(_url) ? _url : {};
+        } = Predicate.isRecord(_url) ? _url : {};
+        const httpClient = (yield* HttpClient.HttpClient).pipe(
+          HttpClient.filterStatusOk,
+        );
 
         const arrayBuffer = yield* HttpClientRequest.get(url).pipe(
           HttpClientRequest.modify({ headers: {} }),
-          HttpClient.filterStatusOk(yield* HttpClient.HttpClient),
-          HttpClientResponse.arrayBuffer,
+          httpClient.execute,
+          Effect.flatMap((_) => _.arrayBuffer),
           Effect.mapError((error) => {
             downloadErrors[idx] = UploadThingError.toObject(
               new UploadThingError({
@@ -121,6 +120,7 @@ export const downloadFiles = (
             );
             return Effect.succeed(undefined);
           }),
+          Effect.scoped,
         );
 
         return new UTFile([arrayBuffer], name, {